# model/src/main/java/com/showtime/model/view/sysuser/SysUserView.java
package com.showtime.model.view.sysuser;

import com.showtime.model.entity.sysuser.SysUser;
import io.swagger.annotations.ApiModel;

import java.io.Serializable;

/**
 * <b><code>SysUserView</code></b>
 * <p/>
 * SysUser
 * <p/>
 * <b>Creation Time:</b> 2017/05/17 21:18.
 *
 * @author lvxin
 * @version 1.0.0
 * @since model 1.0.0
 */
@ApiModel
public class SysUserView extends SysUser implements Serializable {

    /**
     * The constant serialVersionUID.
     */
    private static final long serialVersionUID = 8672084909411877404L;


}